+++ src/main/resources/theme/rciam-group/account/src/app/group-widgets/GroupEnrollment/CreateEnrollmentModal.tsx
@@ -50,7 +50,6 @@ export const EnrollmentModal: FC<any> = (props) => {
     const [enrollment,setEnrollment] = useState<any>({});
     const [isModalOpen,setIsModalOpen] = useState<boolean>(false)
     const [errors,setErrors] = useState<any>({});
-    const [validationRules,setValidationRules] = useState<any>({});
     const [touched,setTouched] = useState<any>(touchDefault);
 
     useEffect(()=>{
@@ -60,6 +59,7 @@ export const EnrollmentModal: FC<any> = (props) => {
         else{
             setIsModalOpen(false);
         }
+        
       setEnrollment({...props.enrollment});
     },[props.enrollment]);
 
@@ -69,18 +69,14 @@ export const EnrollmentModal: FC<any> = (props) => {
         }
     },[enrollment]);
 
-    useEffect(()=>{
-      setValidationRules(props.validationRules);
-    },[props.validationRules]);
-
     
     const validateEnrollment = () => {
         let errors: Record<string, string> = {};
         !(enrollment?.name?.length>0) && (errors.name = Msg.localize('requredFormError'));
         enrollment?.aup?.url?.length>0 && !reg_url.test(enrollment.aup.url) &&  (errors.aup_url = Msg.localize('invalidUrlFormError'));
-        !(enrollment?.aup?.url?.length>0) && validationRules?.aupEntity?.required===true && (errors.aup_url = Msg.localize('requredFormError'));
+        !(enrollment?.aup?.url?.length>0) && props.validationRules?.aupEntity?.required===true && (errors.aup_url = Msg.localize('requredFormError'));
         !(enrollment?.groupRoles?.length>0) && (errors.groupRoles=Msg.localize('groupRolesFormError'));
-        (enrollment?.membershipExpirationDays===0 && validationRules?.membershipExpirationDays?.required===true) && (errors.membershipExpirationDays=Msg.localize('fieldMaxZeroFormError'));
+        (enrollment?.membershipExpirationDays===0 && props.validationRules?.membershipExpirationDays?.required===true) && (errors.membershipExpirationDays=Msg.localize('fieldMaxZeroFormError'));
         (enrollment?.membershipExpirationDays&&!(enrollment?.membershipExpirationDays>0)) && (errors.membershipExpirationDays=Msg.localize('expirationDaysPositiveFormError'));        
         (typeof(enrollment?.membershipExpirationDays)!=='number') && (errors.membershipExpirationDays=Msg.localize('expirationDaysNumberFormError'));
         (enrollment?.commentsNeeded&& (!enrollment?.commentsLabel||enrollment?.commentsLabel.length<1) && (errors.commentsLabel=Msg.localize('requredFormError')));
@@ -94,14 +90,13 @@ export const EnrollmentModal: FC<any> = (props) => {
             !(parsedDate instanceof Date &&isFinite(parsedDate.getTime())) && (errors.validFrom=Msg.localize('validFromInvalidFormError'));
           }
         }
-        if(Object.keys(validationRules).length !== 0){
-          for(const field in validationRules){
-            field==='validFrom'&& validationRules[field]?.required&& !enrollment?.validFrom && (errors.validFrom=Msg.localize('validFromRequiredFormError'));
-            validationRules[field]?.max && parseInt(validationRules[field].max) && (enrollment[field] > parseInt(validationRules[field]?.max )) && (errors[field]=Msg.localize('fieldMaxFormError')+ " (" + validationRules[field]?.max + ")" )   
+        if(Object.keys(props.validationRules).length !== 0){
+          for(const field in props.validationRules){
+            field==='validFrom'&& props.validationRules[field]?.required&& !enrollment?.validFrom && (errors.validFrom=Msg.localize('validFromRequiredFormError'));
+            props.validationRules[field]?.max && parseInt(props.validationRules[field].max) && (enrollment[field] > parseInt(props.validationRules[field]?.max )) && (errors[field]=Msg.localize('fieldMaxFormError')+ " (" + props.validationRules[field]?.max + ")" )   
             // validationRules[field]?.max && parseInt(validationRules[field].max) && enrollment[field]===0 && (errors[field]=  Msg.localize('fieldMaxZeroFormError')+ " (" + validationRules[field]?.max + ")" )
           }
         }
-        
         setErrors(errors);
     }
 
@@ -191,10 +186,18 @@ export const EnrollmentModal: FC<any> = (props) => {
 
       const isPastDate = (date: Date): string => {
         const currentDate = new Date();
-        if (date < currentDate&&(props.enrollment.validFrom!==dateFormat(date))) {
+        
+        // Normalize both dates to remove the time part for an accurate comparison
+        const currentDateWithoutTime = new Date(currentDate.getFullYear(), currentDate.getMonth(), currentDate.getDate());
+        const selectedDateWithoutTime = new Date(date.getFullYear(), date.getMonth(), date.getDate());
+
+        // Check if the selected date is in the past and not the same as the validFrom date
+        if (
+          selectedDateWithoutTime < currentDateWithoutTime &&
+          props.enrollment.validFrom !== dateFormat(selectedDateWithoutTime)
+        ) {
           return Msg.localize('validFromPastFormError');
-        }
-        else{
+        } else {
           return "";
         }
         
@@ -204,7 +207,6 @@ export const EnrollmentModal: FC<any> = (props) => {
     return (
       <React.Fragment>
         <Loading active={loading}/>
-
         <Modal
                 variant={ModalVariant.large}
                 header={
@@ -273,11 +275,9 @@ export const EnrollmentModal: FC<any> = (props) => {
 
                     </Tooltip>
                     ,
-                    
                     <Button key="cancel" variant="link" onClick={()=>{ close()}}>
                         <Msg msgKey='Cancel' />
                     </Button>
-                    
                 ]}
                 >
                     <ConfirmationModal modalInfo={modalInfo}/>
@@ -316,12 +316,11 @@ export const EnrollmentModal: FC<any> = (props) => {
                               }}
                               // helperText=""
                           >
-                          <Tooltip  {...(!(validationRules?.membershipExpirationDays?.required) ? { trigger:'manual', isVisible:false }:{trigger:'mouseenter'})} content={<div><Msg msgKey='enrollmentConfigurationExpirationSwitchDisabledTooltip' /></div>}>
-                                                          
+                          <Tooltip  {...(!(props.validationRules?.membershipExpirationDays?.required) ? { trigger:'manual', isVisible:false }:{trigger:'mouseenter'})} content={<div><Msg msgKey='enrollmentConfigurationExpirationSwitchDisabledTooltip' /></div>}>
                               <Switch
                                 id="simple-switch-membershipExpirationDays"
                                 aria-label="simple-switch-membershipExpirationDays"
-                                isDisabled={validationRules?.membershipExpirationDays?.required&& (isIntegerOrNumericString(enrollment?.membershipExpirationDays)&&enrollment.membershipExpirationDays!==0)}
+                                isDisabled={props.validationRules?.membershipExpirationDays?.required&& (isIntegerOrNumericString(enrollment?.membershipExpirationDays)&&enrollment.membershipExpirationDays!==0)}
                                 isChecked={isIntegerOrNumericString(enrollment?.membershipExpirationDays)&&enrollment.membershipExpirationDays!==0}
                                 onChange={(value)=>{
                                     
@@ -330,12 +329,11 @@ export const EnrollmentModal: FC<any> = (props) => {
                                         setEnrollment({...enrollment});
                                     }
                                     else{
-                                        enrollment.membershipExpirationDays = ((validationRules?.membershipExpirationDays?.defaultValue&& isIntegerOrNumericString(validationRules?.membershipExpirationDays?.defaultValue))?parseInt(validationRules.membershipExpirationDays.defaultValue):3)   
+                                        enrollment.membershipExpirationDays = ((props.validationRules?.membershipExpirationDays?.defaultValue&& isIntegerOrNumericString(props.validationRules?.membershipExpirationDays?.defaultValue))?parseInt(props.validationRules.membershipExpirationDays.defaultValue):32)   
                                         setEnrollment({...enrollment});
                                     }
                                 }}
                                 />
-                            
                             </Tooltip>
                             
                             {enrollment.membershipExpirationDays===0?
@@ -543,14 +541,14 @@ export const EnrollmentModal: FC<any> = (props) => {
                         >
                             <FormGroup
                                 label={Msg.localize('URL')}
-                                isRequired={validationRules?.aupEntity?.required===true}
+                                isRequired={props.validationRules?.aupEntity?.required===true}
                                 fieldId="simple-form-name-01"
                                 helperTextInvalid={touched.aup_url&&errors.aup_url}
                                 validated={errors.aup_url&&touched.aup_url?'error':'default'}
                                 // helperText=""
                             >
                                 <TextInput
-                                isRequired={validationRules?.aupEntity?.required===true}
+                                isRequired={props.validationRules?.aupEntity?.required===true}
                                 type="url"
                                 id="simple-form-name-01"
                                 name="simple-form-name-01"